requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
yfinance==0.2.28
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
matplotlib==3.7.2
python-dotenv==1.0.0
